Rebase the remaining UI, chrome and theming onto DOMjudge 9.0.1
Phase 7, and the last of the fork's 261 touched files.

Method: 30 files that upstream never touched between 8.3.0 and 9.0.1 were taken
from the fork wholesale (the menus, the theme switcher, the dark-mode
bootstrap, the date-time picker, diff_match_patch, favicons, robots.txt and
several forms and partials). The remaining 37 were applied as per-file
three-way merges; 21 went in clean and 16 needed hand resolution, 18 conflict
hunks in total. Notable resolutions:

- domjudge.js keyboard shortcuts: kept upstream's modifier-key guard and the
  fork's exclusion for the Editor.js blog editor; both are wanted.
- ScoreboardMatrixItem: kept upstream's named-argument signature with its
  string $points, which is what the scoring work already feeds it. The fork's
  float variant would have broken upstream's callers.
- SingleTeamScoreboard now passes points too, so a team's own view shows the
  same partial score as the public board.
- jury/contests.html.twig: took the fork's rewritten accordion, then re-added
  upstream's contest-detail link, which the rewrite had dropped.
- jury/problem.html.twig: kept both attachment columns. Upstream added
  "Mime type"; the fork uses "Type" for the editor language, so neither
  replaces the other.
- team submission list: kept upstream's `>=` endtime comparison, which is a
  fix the fork predates - a submission exactly at endtime is outside the
  contest.
- jury/base.html.twig: the fork's configurable website_title, with upstream's
  'DOMjudge' kept as the fallback so an unset title still renders.
- config_check.html.twig: took the fork's data-bs-parent; upstream still had
  the Bootstrap 4 spelling.
- Added the printPartialPointsScoringMode filter, deferred out of phase 3 and
  needed here by the jury contest page.

Verified against the running stack: 205 Twig templates parse, the container
lints, every literal route reference in templates and src resolves, every
changed PHP file parses, doctrine:schema:validate reports no new mapping
errors, and all 29 public, team and jury pages return 200.

Not verified: appearance. Dark mode, the responsive work and the redesigned
9.0 pages need a browser and your judgement.

+++ webapp/src/Utils/Scoreboard/Summary.php
@@ -6,7 +6,7 @@ use App\Entity\ContestProblem;
 
 class Summary
 {
-    /** @var int[] */
+    /** @var float[] */
     protected array $numberOfPoints = [];
 
     /** @var int[] */
@@ -28,12 +28,12 @@ class Summary
         }
     }
 
-    public function getNumberOfPoints(int $sortorder): int
+    public function getNumberOfPoints(int $sortorder): float
     {
         return $this->numberOfPoints[$sortorder] ?? 0;
     }
 
-    public function addNumberOfPoints(int $sortorder, int $numberOfPoints): void
+    public function addNumberOfPoints(int $sortorder, float $numberOfPoints): void
     {
         if (!isset($this->numberOfPoints[$sortorder])) {
             $this->numberOfPoints[$sortorder] = 0;
